Fix overflow with small playback durations.

+++ src/lib/player/Player.c
@@ -449,7 +449,7 @@ int64_t Player_get_nanoseconds(const Player* player)
 {
     assert(player != NULL);
 
-    const int32_t ns_this_audio_rate =
+    const int64_t ns_this_audio_rate =
         player->audio_frames_processed * 1000000000LL / player->audio_rate;
     return player->nanoseconds_history + ns_this_audio_rate;
 }
